v14 WS-H/T20: whisper_target segmenter extractor + fp32 teacher-build fix
Chunk A of the BT full-run prep. Wires the P3 Whisper teacher target end to
end and fixes the teacher-cache build crash on transformers v5.

whisper_target extractor (new):
- WhisperTargetExtractor(BaseStatic): aggregation="trigger", slices the cached
  whole-movie Whisper stream at the trigger Word's MOVIE-clock onset
  (movie_onset_s), returns (250, 1280) CPU float32; mmap-loads + memoizes the
  dense stream per movie; clamps the tail window to exactly 250 frames.
- FLAG 9: keys off movie_onset_s, NOT the Δlag-shifted neural segmenter start.
  The neural vs movie clocks diverge 235-904 s within one BT trial; keying off
  the neural clock would slice the wrong 5 s of the movie. 14 unit tests,
  including the clock-divergence regression.

word_events.py: thread movie_onset_s (= transcript words_df['start'], the movie
clock) onto every emitted Word/nonverbal row so the extractor can join on it.

dispatch_v14.py: whisper_target_cache_dir / whisper_layer_merge params; when set
(P3), insert WhisperTargetExtractor into the segmenter (no _apply_extractor_cache
— the whole-movie cache IS the precompute). Rewrote the WS-H comment that wrongly
told a future implementer to anchor the teacher to est_idx/2048 (the exact FLAG 9
inversion). 3 wiring tests.

teacher-cache fp32 fix (the build-blocking bug): transformers v5 loads large-v3
in its native fp16, so the encoder conv1d saw Half weights vs a float32 mel and
raised "Input type (float) and bias type (c10::Half)". extract() now casts the
mel to the model's param dtype (correctness, both dtypes); the build forces fp32
(fidelity — the mean over 32 encoder layers must match the fp32 ceiling-probe
that validated mean_all; ≈6 GB, fits the 32 GB RTX 5000 Ada). dtype-alignment
regression test added.

Full suite 1121 passed, 5 skipped.

+++ scripts/neuroprobe/build_bt_teacher_cache.py
@@ -174,7 +174,15 @@ def main(argv: list[str] | None = None) -> int:
     if device == "cuda" and not torch.cuda.is_available():
         print("  WARNING: cuda requested but unavailable; falling back to cpu.")
         device = "cpu"
-    model = WhisperForConditionalGeneration.from_pretrained(args.model).eval().to(device)
+    # Force fp32. transformers v5 loads a checkpoint in its NATIVE dtype, and
+    # large-v3's HF weights are fp16 — but the teacher target is a mean over all
+    # 32 encoder layers, and accumulating that in fp16 drifts from the fp32
+    # ceiling-probe that validated `mean_all`. fp32 large-v3 ≈6 GB, well within
+    # the 32 GB RTX 5000 Ada; the dense stream is still stored fp16 downstream.
+    model = (
+        WhisperForConditionalGeneration.from_pretrained(args.model)
+        .eval().float().to(device)
+    )
     proc = WhisperProcessor.from_pretrained(args.model)
     fe = WhisperFeatureExtractor(model, proc, layer_merge=merge)
